Finish Moving
Finished moving Settings features into the main layout side bar
