+++ resources/js/Pages/Settings/Index.vue
@@ -1,10 +1,8 @@
 <template>
-    <h1 class="text-overline d-flex align-center mx-2 mb-2" style="gap: 0.3rem">
-        <v-icon>mdi-cog-outline</v-icon>
-        <p>Settings</p>
-    </h1>
+    <Heading icon="mdi-cog-outline">Settings</Heading>
     <v-card>
-        <v-tabs v-model="tab.tab" color="primary">
+        No Content Yet.
+        <!-- <v-tabs v-model="tab.tab" color="primary">
             <v-tab
                 v-for="(tab, index) in tab.tabData"
                 :value="index"
@@ -15,52 +13,23 @@
                 >{{ tab.title }}</v-tab
             >
         </v-tabs>
-        <v-container>
-            <router-view v-slot="{ Component }">
-                <template v-if="Component">
-                    <Suspense timeout="0">
-                        <component :is="Component" />
-
-                        <template v-slot:fallback>
-                            <Spinner />
-                        </template>
-                    </Suspense>
-                </template>
-            </router-view>
-        </v-container>
+        <v-container> No Content </v-container> -->
     </v-card>
 </template>
 
 <script>
-import Spinner from "@/components/Spinner.vue";
+import Heading from "../../components/Heading.vue";
+
 export default {
-    components: {
-        Spinner,
-    },
     data() {
         return {
             tab: {
                 tab: null,
-                tabData: [
-                    {
-                        icon: "mdi-shield-outline",
-                        title: "Disciplinary Measures",
-                        link: { name: "disciplinaryMeasure" },
-                    },
-                    {
-                        icon: "mdi-format-list-checkbox",
-                        title: "Violation Types",
-                        link: { name: "violationType" },
-                    },
-                    {
-                        icon: "mdi-alert-decagram-outline",
-                        title: "Violations",
-                        link: { name: "violation" },
-                    },
-                ],
+                tabData: [],
             },
         };
     },
+    components: { Heading },
 };
 </script>
