+++ resources/js/router/index.js
@@ -33,34 +33,23 @@ const routes = [
             {
                 path: "settings",
                 name: "settings",
-                redirect: { name: "disciplinaryMeasure" },
                 component: () => import("@/Pages/Settings/Index.vue"),
-                children: [
-                    {
-                        path: "disciplinary-measure",
-                        name: "disciplinaryMeasure",
-                        component: () =>
-                            import(
-                                "@/Pages/Settings/TabContents/DisciplinaryMeasures.vue"
-                            ),
-                    },
-                    {
-                        path: "violation-type",
-                        name: "violationType",
-                        component: () =>
-                            import(
-                                "@/Pages/Settings/TabContents/ViolationType.vue"
-                            ),
-                    },
-                    {
-                        path: "violation",
-                        name: "violation",
-                        component: () =>
-                            import(
-                                "@/Pages/Settings/TabContents/Violation.vue"
-                            ),
-                    },
-                ],
+            },
+            {
+                path: "disciplinary-measure",
+                name: "disciplinaryMeasure",
+                component: () =>
+                    import("@/Pages/DisciplinaryMeasures/Index.vue"),
+            },
+            {
+                path: "violation-type",
+                name: "violationType",
+                component: () => import("@/Pages/Violation/ViolationType.vue"),
+            },
+            {
+                path: "violation",
+                name: "violation",
+                component: () => import("@/Pages/Violation/Index.vue"),
             },
             {
                 path: ":showable+/:id/show",
